Add can mock
