+++ app/src/machine/can_master.py
@@ -1,142 +1,13 @@
 from enum import IntEnum
 from typing import Any
 import can
-import struct
 from time import sleep
 import datetime
 
-
-class CanTimeoutException(Exception):
-    pass
+from src.machine.can_master_base import Battery, CanInfo, CanMasterBase, OilPress, OilTemp, Rpm
 
 
-class RpmStatus(IntEnum):
-    LOW = 0
-    MIDDLE = 1
-    HIGH = 2
-
-
-class Rpm(int):
-    LOW_THRESHOLD = 3200
-    HIGH_THRESHOLD = 6400
-    MAX = 9600
-
-    @property
-    def status(self) -> RpmStatus:
-        if self < self.LOW_THRESHOLD:
-            return RpmStatus.LOW
-        elif self < self.HIGH_THRESHOLD:
-            return RpmStatus.MIDDLE
-        else:
-            return RpmStatus.HIGH
-
-
-class WaterTempStatus(IntEnum):
-    LOW = 0
-    MIDDLE = 1
-    HIGH = 2
-
-
-class WaterTemp(int):
-    LOW_THRESHOLD = 50
-    HIGH_THRESHOLD = 100
-
-    @property
-    def status(self) -> WaterTempStatus:
-        if self < self.LOW_THRESHOLD:
-            return WaterTempStatus.LOW
-        elif self < self.HIGH_THRESHOLD:
-            return WaterTempStatus.MIDDLE
-        else:
-            return WaterTempStatus.HIGH
-
-
-class OilTempStatus(IntEnum):
-    LOW = 0
-    MIDDLE = 1
-    HIGH = 2
-
-
-class OilTemp(int):
-    LOW_THRESHOLD = 50
-    HIGH_THRESHOLD = 120
-
-    @property
-    def status(self) -> OilTempStatus:
-        if self < self.LOW_THRESHOLD:
-            return OilTempStatus.LOW
-        elif self < self.HIGH_THRESHOLD:
-            return OilTempStatus.MIDDLE
-        else:
-            return OilTempStatus.HIGH
-
-
-class OilPressStatus(IntEnum):
-    LOW = 0
-    HIGH = 1
-
-
-class OilPress(float):
-    THRESHOLD = 3.0
-
-    @property
-    def status(self) -> OilPressStatus:
-        if self < self.THRESHOLD:
-            return OilPressStatus.LOW
-        else:
-            return OilPressStatus.HIGH
-
-
-class FuelRemainStatus(IntEnum):
-    LOW = 0
-    HIGH = 1
-
-
-class FuelRemain(float):
-    THRESHOLD = 1.0
-
-    @property
-    def status(self) -> FuelRemainStatus:
-        if self < self.THRESHOLD:
-            return FuelRemainStatus.LOW
-        else:
-            return FuelRemainStatus.HIGH
-
-
-class BatteryStatus(IntEnum):
-    LOW = 0
-    HIGH = 1
-
-
-class Battery(float):
-    THRESHOLD = 11.0
-
-    @property
-    def status(self) -> BatteryStatus:
-        if self < self.THRESHOLD:
-            return BatteryStatus.LOW
-        else:
-            return BatteryStatus.HIGH
-
-
-class LapTime(datetime.timedelta):
-    pass
-
-
-class CanInfo:
-    rpm: Rpm
-    oilTemp: OilTemp
-    oilPress: OilPress
-    battery: Battery
-
-    def __init__(self) -> None:
-        self.rpm = Rpm(0)
-        self.oilTemp = OilTemp(0.0)
-        self.oilPress = OilPress(0)
-        self.battery = Battery(0.0)
-
-
-class CanMaster:
+class CanMaster(CanMasterBase):
 
     canInfo: CanInfo
 
@@ -170,7 +41,7 @@ class CanMaster:
 
     #     # self.canInfo.rpm = 3333  # TODO fix
 
-    def receiveData(self) -> bytearray:
+    def _receiveData(self) -> bytearray:
         values = bytearray(b'')
 
         retryLimit = 12
@@ -184,7 +55,7 @@ class CanMaster:
         return values
 
     def updateCanInfo(self):
-        data = self.receiveData()
+        data = self._receiveData()
         self.canInfo.rpm = Rpm(data[CanMaster.DBS_RPM[0]] * 256 +
                                data[CanMaster.DBS_RPM[1]])
         self.canInfo.oilTemp = OilTemp(
